Finish Fri Fall 12
Note that the weekly schedule as it stands is like this:
- Tuesday: Casks day (cellar)
- Friday: Ginger Island day (ancient fruits planted one week apart starting Fall 5, and grow in 28 days)
- Saturday: Keg day (quarry)

We should soon synchronize the ancient fruits in the greenhouse, once
all of them grow up.  They would ideally be harvested on Keg day so
that we can harvest and then immediately move them to kegs.  However since we have a backlog of starfruit this is unnecessary for at least

Farm:
- One cranberry harvest, 70000g
- Blackberry harvest, 93000g (140 + 250 + 100 + 200 = 712 blackberries)
- One Cheese harvest, about 30000g (30 iridium cheese)
- Autoharvesters are great
- Added some bee houses
- Took 100 Grapes quest and planted 14 + 5 more grapes
- Built more truffle oil machines.
- Finished Lobster quest

Ginger Island:
- Planted 84 more ancient fruit, about 70 fertilized
- Planted fairy rose in left side
- Finished Pirate Wife quest
- Enchanted sword: Crusader
- Volcano dungeon not much gained

TODO:

Ginger Island:
- Fairy rose on Ginger Island grows up on Fall 24, so transfer all bee
  houses there at that time or at end of fall.
- Find cinder shards for more enchantments (esp pickaxe)
- Find dragon teeth for island obelisk.  Or should we get the beach obelisk?

Farm:
- Buy more pigs about 10 days before end of winter (probably sell goats)
- Build more tappers (need copper)
- Build more bee houses (need coal [from dust spirits in mines?])
- Build more kegs (need copper [and iron])

Other:
- Sandy's birthday 15th--go to skull caverns? (Very low on ores esp copper)
- Could get copper from normal mines
- Stardew Valley Fair 16th

At end of winter, let's redo the 4x3x3 plot as a full-sized 5x5 plot
and take the slight losses.  It's not worth the trouble.